Rename efficientnet_b0 to efficientnet
