+++ src/otx/recipe/classification/h_label_cls/efficientnet_b0.yaml
@@ -1,5 +1,5 @@
 model:
-  class_path: otx.algo.classification.efficientnet_b0.EfficientNetForHLabelCls
+  class_path: otx.algo.classification.efficientnet.EfficientNetForHLabelCls
   init_args:
     version: b0
