+++ tests/integration/api/test_engine_api.py
@@ -7,7 +7,7 @@ from pathlib import Path
 
 import pytest
 from openvino.model_api.tilers import Tiler
-from otx.algo.classification.efficientnet_b0 import EfficientNetB0ForMulticlassCls
+from otx.algo.classification.efficientnet import EfficientNetForMulticlassCls
 from otx.core.config.hpo import HpoConfig
 from otx.core.data.module import OTXDataModule
 from otx.core.model.base import OTXModel
@@ -184,7 +184,7 @@ def test_otx_hpo(
         reason = f"test_otx_hpo for {task} isn't prepared yet."
         pytest.xfail(reason=reason)
 
-    model = EfficientNetB0ForMulticlassCls(label_info=2)
+    model = EfficientNetForMulticlassCls(label_info=2)
     hpo_config = HpoConfig(metric_name=METRIC_NAME[task], expected_time_ratio=2, num_workers=1)
     work_dir = str(tmp_path)
     engine = Engine(
